Rename schnorr to oldschnorr and add fixes to build system
