+++ Makefile.am
@@ -141,8 +141,8 @@ if ENABLE_MODULE_MULTISET
 include src/modules/multiset/Makefile.am.include
 endif
 
-if ENABLE_MODULE_SCHNORR
-include src/modules/schnorr/Makefile.am.include
+if ENABLE_MODULE_OLDSCHNORR
+include src/modules/oldschnorr/Makefile.am.include
 endif
 
 if ENABLE_MODULE_RECOVERY
